MNT Refactor eagerloading tests into their own class
